refactor redundant files
